# contracts/auth.openapi.yaml
openapi: 3.1.0
info:
  title: Carelink Auth API
  version: 1.0.0
  description: |
    Auth and onboarding REST API for Carelink mobile client.
    Full spec: docs/backend/auth/API.md
servers:
  - url: https://api.carelink.app/v1
    description: Production (v1 — pairing, profile, tokens)
  - url: https://api.staging.carelink.app/v1
    description: Staging (v1)
  - url: https://api.carelink.app/v2
    description: Production (v2 — workspace email auth)
  - url: https://api.staging.carelink.app/v2
    description: Staging (v2 — workspace email auth)
...
paths:
...
  /auth/workspace/sign-in/delegate-status:
    get:
      tags: [workspace-auth]
      operationId: workspaceSignInDelegateStatus
      summary: Waiting member polls delegation outcome
      parameters:
        - name: challengeId
          in: query
          required: true
          schema:
            type: string
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: pending | approved | rejected | expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceDelegateStatusResponse'
...
components:
  securitySchemes:
    userRefreshAuth:
      type: apiKey
      in: header
      name: X-User-Refresh
      description: Opaque user-scoped refresh token from phone register or sign-in (create-family path).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceToken:
      type: apiKey
      in: header
      name: X-Service-Token
      description: Internal service token (INTERNAL_SERVICE_TOKEN) for keeper/backend reject.
...
  schemas:
...
    WorkspaceDelegateStatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        deviceSession:
          type: string
          description: Present only for the waiting member when status is approved
...